SXSentenceSelector: Wait for MT providers to be initialized
When "Pick a sentence" step is loaded and SXSentenceSelector SFC
is mounted, a vuex action is dispatched to select the section title
as translation segment. However, this action needs current MT
provider to be set in order to work properly. This used to work
properly but during refactoring an "await" for initializeMTProviders
action was dropped, leading to a bug.

Restore the "await" for initializeMTProviders vuex action, inside
onMounted callback in SXSentenceSelector SFC, so that the
currentMTProvider state variable is properly set before translating
the first segment.

Change-Id: Ia3a6e7e63f1f88bd078012c644df4fdf22594028