SX page contents: resolve references asynchronously
When user visits "Compare contents" step, there is a long waiting
time until the section contents are loaded. This is primarily caused
because of the long execution time of the reference resolving
functionality. However, these references do not need to be resolved
until the user starts the actual translation (in "Pick a sentence"
step), so we do not need to wait for it to be completed.

In order to reduce the waiting time:

* Remove fetchPageSections vuex action and api method, and use
  fetchSegmentedContent api method to fetch the page contents,
  without resolved references.

* Add resolvePageContentReferences vuex action that uses the page
  contents as they are stored inside, and updates the page sections
  to include resolved references. This action is being dispatched
  asynchronously, and we do not wait for it to be completed.

Bug: T290453
Change-Id: I55c2f794810c9e0d91e7c627095d1c6ccd8faf57