SX language selector entry point: Navigate to SX within the same tab
Until now, every time the buttons for this entrypoint are clicked,
a new window opens, to navigate the user to SX application. However,
it is desired for this navigation to happen within the same tab, and
not by opening new tab/window.

Fix this by removing the 'target="_blank"' attribute for both buttons
that are provided for this entrypoint.

Bug: T292828
Change-Id: I97e32ba0caeff7e6059db2168dfeb0ed5a5e975f