SXContentComparator: Call hasSectionTranslations getter as a function
vuex "translator/hasSectionTranslations" getter is returning a method
(method-style access), although it doesn't take any argument. Thus,
when it is accessed in property-style, inside SXContentComparator, it
doesn't return a value.

Fix this, temporarily, by access this getter as a method. Getters
should be turned to property-style getters in a separate patch.

Change-Id: I5bac356e72ca2b91471e36e2651d18d05f136e59