SX useApplicationState composable: Fix computed autonym properties
sourceLanguageAutonym/targetLanguageAutonym uses
currentSectionSuggestion state variable to get the related
source and target languages. When currentSectionSuggestion
state variable is not null, its languages are, indeed, the
same as state source/targetLanguage. However,
currentSectionSuggestion is not guaranteed to be not null
at all, so it would be better to use sourceLanguage and
targetLanguage state variables, instead.

Change-Id: I28289d4068ec209fdf17b65f64112c12fcfa790a