SX article search: Show loading indicator as soon as the user types
Currently, the loading indicator for search results is being shown,
when the debounced search function runs for the first time. That is,
500ms after the user has stopped typing their query.

This behaviour leads to a "No pages found" message when no previous
search results exist. When previous results do exist, they are shown
until the user stops typing.

Fix this by activating searchResultsLoading flag, the first time the
user types and deactivating it when the debounced search method is
completed. Following this change, the loading indicator is displayed
as soon as the user starts typing, and until the search results have
been loaded.

Bug: T287835
Bug: T290457
Change-Id: I0624f1406bafb7446cf8c2e135099176de973ccc