SXTranslationConfirmer: Go to previous route on back button click
Currently, when clicking the close icon inside the header of
confirmation step, the user is navigated back to the dashboard,
regardless if the user reaches the confirmation step through the
dashboard or the search step. However, the close icon should lead
to the previous step, dashboard or search.

To implement that, we add a new param named "previousRoute" that
contains the name of the route before to the confirmation step.
This approach was selected because vue router navigation guards
do not work well with vue composition api (for vue-router@3), and
thus we could not get the name of the previous route elseway.
This param is used to navigate to the appropriate route each time
close icon is clicked.

Bug: T290455
Change-Id: I88923a4aac65d3b32f4546c574d41e88b5cc97f7