SXTranslationConfirmer: Clear prefilled section from URL if invalid
Currently prefilled section is populated in "Confirm translation"
step, when "section" URL parameter exists, even if this section
title doesn't exist in the current page. That leads to misleading
information being presented to the user.

Fix this by checking if the section title exists inside the page,
and clearing it from the URL if it doesn't.

Bug: T287742
Change-Id: I2c6c666ae5195df745ab1856e8c1c085bd6fee34