MWMessage: Apply the appropriate color only to icon inside slot
Currently, the color that corresponds to the type of the message,
is applied to all icons inside this message. For example, in case
of "error" type, red50 color is applied to all icons inside this
message. However, only the icon inside the "icon" slot should have
this color.

Fix this by adding the color rule for .mw-ui-message__icon class
instead of .mw-ui-icon class.

Change-Id: Iccf15982a0085cda8fd9227065a6cdafe91ebdcd