Change "New pages" message to "Create new pages"
This message appears as a section title on
the SX dashboard, and another section
on the same page is called
"Expand with new sections".

It makes sense to make both sections sound
like a call to action, with a verb
and a complement.

Without this, it may look like these are
new pages that were recently created on a wiki.

Change-Id: Iafb4caae37abd54189f3294f0b1585b55226f803