SX: Add UI for entrypoint for reviewing recently translated pages
This patches adds the UI for the entrypoint that encourages users
to review recently translated articles.

This UI basically consists of 2 parts: the banner on the top of the
article and the dialog that opens when user clicks on the banner.

The banner for this entrypoint should only be visible when no other
notices (hatnotes, amboxes) exist inside the lead section of this
article. For this reason, this module checks if such notes exist,
before any other action. Additionally, the banner is created and
appended inside the article, using plain Javascript code, so that
this part that is executed as soon the page loads can be as fast as
possible.

However, the dialog UI is implemented inside a Vue Single File
Component, so that the implementation is (much) simpler. Vue and
the component itself are only imported on demand, once user clicks
on the banner.

Bug: T287236
Change-Id: I961d310869d8f65749d09351a0ae3552214a5134