CX: Add "cx_significant_edits" table to support entrypoint
In order to support the new entrypoint that invites users to
translate a section they have previously edited in another language,
a new table needs to be created.

This table stores recent significant edits done by a user in any
wiki, with a maximum of 10 edits stored per user. For each edit
the user id, the page wikidata id, the language, the title of the
page and the section titles that were affected are stored.

Abstract schema also added.

Bug: T287025
Change-Id: I4fb6e6d1b5c05a59806c204221c698485f5a3ff9