SX Publish: Remove MT validation from publishTranslation action
MT validation, currently happens as soon as the user visits the
"Preview and publish" step. Thus, user is aware about possible
MT warnings before publishing a translation. Additionally, the
"Publish" button is disabled when MT errors exist, so it's
impossible to publish a translation with MT errors.

As a result, MT validation is no longer needed inside
"publishTranslation" vuex action and should be removed as it
blocks translations with MT warnings to be published.

Bug: T260134
Change-Id: Iebf9861828add6f074ce37c1f7462c20e416c9e4