SX Refactoring: Abstract initializeLanguages in separate composable
useDashboardInitialization composable is used inside Dashboard, but
also inside SXArticleSearch. However, even the naming of the
composable indicates that it doesn't exactly belong inside
SXArticleSearch SFC. In fact, SXArticleSearch only needs the
language initialization and the translation fetching.

To fix this issue, abstract initializeLanguages in a separate
composable, so that it can be used inside SXArticleSearch, as it
should be.

useDashboardInitialization will be moved to components/CXDashboard
directory in a following patch.

Change-Id: I600f53213455ede3460362ad25c008418fbba5de