SX recent edit entrypoint: add hook to register JS module
In order to support the new entrypoint to invites users to translate
a section they recently edited in another language, a new hook handler
for the "BeforePageDisplay" event should be added.

This patch adds a hook that registers 'ext.cx.entrypoints.recentedit'
JS module, when the appropriate conditions are met and also exposes
a Javascript variable containing the edited sections, the respective
language and the respective page title of each related edit.

It also adds "findEditsForPotentialSuggestions" method inside store
for RecentSignificantEdit model to support fetching related edits
from the database.

Finally, it adds an empty 'ext.cx.entrypoints.recentedit' JS module
inside extension.json. The actual implementation of this module will
be added in the following patch.

Bug: T287025
Change-Id: Ifc9220b1bc4cf5718124bb0ee208917de058bf7c