CX recent edit entrypoint: Add edit model and store
In order to support the new entrypoint to invite users to translate
a section they recently edited in another language, we should add
a new model to represent such significant edits.

This patch adds RecentSignificantEdit, a value object class to
represent these edits and also adds RecentSignificantEditStore,
a store class to handle all database interactions concerning
this model. The functions that are added in this patch will be
used inside a hook to properly create new significant edits as
needed.

This patch also registers RecentSignificantEditStore as a service,
inside ServiceWiring.php file, so that Dependency Injection is
properly implemented for this class.

Bug: T287025
Change-Id: I95e16c455045e68dd5578cb8722939a3074654e1