SX: Add after edit hook to create entries in significant-edits table
In order to support the new entrypoint to invite users to translate
a section they recently edited in another language, we should add the
information for the significant edits, that affect specific sections,
to the "significant-edits" SQL table.

To support this need, this patch creates a RevisionFromEditComplete
hook handler, where the new entries for the significant edits are
created.

Bug: T287025
Change-Id: I09c95370476d32dab9f2d776625df0448c146370