SX: Refactor vuex translation functionality
Until now, action, getters and mutations related to translation
were duplicated for section titles and section sentences. That
led to a lot of redundant code that decreased code readability
and flexibility.

In addition to that, upcoming patches will add support for block
templates, that will bring even more complexity in these vuex
files.

In order to avoid piling up such complexities that will
end up to a spaghetti code, and facilitate the addition of new
relevant functionality to support block templates, this patch:

* Notation for translation "segment" is replaced by "translation
  unit" as it seems more generic, and it will also be used for
  subsections that represent block templates in following patches.
* Merges the duplicate vuex actions for translating section
  titles and sentences into single actions.
  e.g. "translateSectionTitle", "translateFollowingSentence",
  "translateSelectedSegment" and "translateSelectedSentence"
  actions are merged to "translateTranslationUnitById".
* Getters related to currently selected translation are moved
  to pageSection model as all related information is already
  part of this model, and it's easier to add extra getters there.
* Fix components that use these actions, getters or mutations
  to use the new versions of them.
* Fix tests too.

Bug: T284058
Change-Id: I36d57798aad9416c60719e8998c7da9316536b4e