SX: Support block template concept inside vuex and pageSection model
In order to support block template adaptation inside "Pick a
sentence" screen, we have to properly support it inside vuex store
and inside pageSection and subSection models.

To this cause, this patch adds the required properties and getters
inside subSection model (blockTemplateSelected,
blockTemplateTranslatedContent, blockTemplateProposedTranslations,
isBlockTemplate, etc) that are used to select a subsection as a
block template, and store the proposed or applied translations for
these subsections.

This patch also adds all required adjustments inside vuex actions,
mutations and getters, and pageSection model to expand the concept
of "sentence" to "translation units" that support both sentences
and block templates.

SFCs and tests fixed too.

UI to support block templates inside "Pick a sentence" step will
be added in following patch.

Bug: T294650
Change-Id: Id84f96853269b0ddf895713c876a2776114c9df5