SX translator api: Add parseTemplateWikitext method
When translating a block template, the translation from cxserver
contains the translated/adapted template params but not the HTML
of the template itself. In order to successfully translate a block
template (e.g. infobox), we need to parse the wikitext of the
template translation from cxserver, to get the proper HTML for
rendering the template.

This patch adds "parseTemplateWikitext" to support this need. This
method returns a promise that resolves to the template HTML, given
the template wikitext, the page title and the page language.

Bug: T294650
Change-Id: Ie0e7eaa509931ad96286e1bed24949b29e3a9da6