SX vuex translateTranslationUnitById action: support block templates
application/translateTranslationUnitById action is the action that
is used to translate all kind of translation units (section title,
section sentences and block templates). However, block template
translation/adaptation is not supported until now.

To support translation/adaptation of block templates, this patch
adds the related logic inside this vuex action, by parsing the
translation returned by cxserver, to get the HTML for rendering the
template. While at it, also add check if the proposed translation
already exists for the specific MT provider so that we can avoid
unnecessary requests.

Bug: T294650
Change-Id: If701e62c497e29e87af43500dacfc54271a18613