SX: Fix suggestion loading after canceling preselected translation
When user visits a SX URL with preselected page title, they are
navigated to the "Confirm translation" step. In this case, if the
user clicks on the "close" icon, they are navigated back to the
dashboard screen but page and section suggestions are not being
filled, and only the preselected suggestion is shown inside the
section suggestions list (even when it is a page suggestion).

The reason why suggestions are not fetched is that
"suggestions/initializeSuggestions" vuex action, returns without
fetching anything when at least one section or page suggestion
exists. Additionally, the suggestion is being shown in section
suggestions list (regardless if it's a section or a page suggestion),
because of a wrong logic inside "suggestions/loadSectionSuggestion"
vuex action.

To fix this issue, this patch:
* Fixes "initializeSuggestions" action to return without fetching,
  only if both page and section suggestion slices are full.
* Fixes "loadSectionSuggestion" action to add the suggestion to
  page suggestions if no section suggestion exists for the given
  title and language pair.

Bug: T296624
Change-Id: I708458f75be6b5ff5783e34240a97b0bfae8e0c9