SX: Fix title route param for navigation to SX Editor
Navigation to SXEditor screen can be accomplished inside "Pick a
sentence" step and inside "Preview and publish" step.

In the first case, the target title is passed as title route param.
However, when translating a lead section the target page title is
missing and thus the title param is undefined, leading to issues
in template parsing and a vue warning about an undefined title
prop for VisualEditor SFC.

In the second case, the title route param is totally missing.

This patch fix above issues appropriately.

Change-Id: Icdd1b7befafb340616bfaaba22d0218c1da52fdd