SX: Check if target template exists before re-rendering
Currently, the proposed translation returned by cxserver for a block
template, is always being parsed to get the HTML for rendering the
target template. However, an equivalent target template is often not
available, and thus this processing is useless.

This patch adds a check whether the target template exists before
this parsing, in order to avoid a useless request, and sets the
proposed translation to null if it doesn't exist. This way we also
make sure that the proposed translation will always be null when the
equivalent template doesn't exist in the target language.

Bug: T295973
Change-Id: I9e4b199c6bd91f8626508c8d0d811718df8d9006