SX: Disable auto-scroll when sentence already inside the viewport
Currently, auto-scrolling acts also on sentences that are completely
within the viewport, and as a result when tapping on one of those
sentences, the viewport moves in an unexpected way.

This patch disables auto-scrolling for such sentences, so that it
works only for sentences that are (at least) partially outside the
viewport.

Concerning these partially hidden sentences, there exists another
minor issue that occurs from time to time: since the scrolling can be
completed before the fetching of the suggested translation, the
proposed translation card at the bottom of the screen can partially
hide the selected sentence, when the proposed translation is fetched
and the card height changes. To fix this issue, this patch changes
the scrolling behaviour to align the sentence at the vertical center,
and avoid such cases.

Screencast after fixes:
https://imgur.com/a/lu9Ih8z

Bug: T297005
Change-Id: Ia2e68f149de0c237469a026754579bcc6f12e6b2