SX: Show invite for SX after publishing in the Sandbox
Until now, the invitation for the user to translate another section,
is only being shown for content pages (pages in a content namespace).
However, this invite should also be shown in sandbox pages, where
user is redirected after successful publishing to sandbox.

This patch adds the "sx.publishing.followup" module also for sandbox
pages. It also conditionally fetches the section suggestions and
displays the last paragraph of the invite, only for content pages,
as these are not needed for the sandbox pages per design.

Bug: T275004
Change-Id: I6ce18044c9a78602361d42735cb59dba84e5e99b