SX suggestions api: Remove slice in fetchPageSuggestions method
fetchPageSuggestions method takes a "count" argument that defines
how many suggestions should be fetched. However, until recently,
the Recommendation API didn't limit the results based on this
parameter. To mock this behaviour, the returned results were
sliced, so that their length is equal to "count" at most.

Since this issue has been resolved for the Recommendation API
(tracked in T287227), the slice is redundant and can be removed.

Change-Id: I70636ba19fe943486b1b4db83b37cd73191ffba0