SX store: Add application/isSandboxTarget getter
Inside Section Translation application, there are several cases that
we need to check if the publishing target is the user's sandbox,
since some functionalities need special handling for this case of
publishing.

This patch adds "application/isSandboxTarget" getter and replaces
the comparisons of "publishTarget" state variable with a plain
string with it.

Change-Id: I3b77d931164ebdfa534c17883eadd9e99666617e