SX: Avoid wikibase link request for sandbox publishing
When a new article is successfully created using Section Translation,
a request to link this article in wikidata is being sent. However,
this should not happen when publishing target is the user's sandbox.

This patch adds a check to avoid this request when publishing to
sandbox.

It also adds a catch block to catch any errors of the wikibase link
request, since such failure is acceptable.

Bug: T297804
Change-Id: If10b368e1ea4598a2ae73329ea4dd33645f83eb5