SX Entrypoints: Add mffrequentlanguages module and register it to RL
This patch adds the JS module that listens to a hook fired when
mobile language selector is rendered, and given the LanguageSearcher
instance, it calculates the missing, frequently used languages that
are enabled for SX. If such languages exist, an HTML div element is
created that contains all the required markup for the entrypoint
banner. The outer HTML of this element is being passed as an argument
to the "addBanner" method, which is a member of the LanguageSearcher
(MobileFrontend extension).

This patch also adds the "BeforePageDisplay" hook handler that
registers the JS module above to the ResourceLoader, when the
appropriate conditions are met.

Bug: T298032
Change-Id: I6321be12caeb9de0eec80c317ae09fd32f606197