CX: Rename Database class to LoadBalancer and register it as service
According to the related Mediawiki documentation, the service locator
pattern should be avoided. When a service needs access to another
service, it should ask for it in the constructor (actual dependency
injection). Documentation here:
https://www.mediawiki.org/wiki/Dependency_Injection

The ContentTranslation/Database class is a static factory using such
a hidden dependency on the load balancer factory. In order to follow
these best practices and use actual dependency injection, we should
refactor it to accept this dependency in the constructor and properly
register it as a custom service inside ServiceWiring.php file. This
way the Database service can be injected in the constructor of other
classes that need it.

This patch renames the Database class to LoadBalancer and refactors
it to follow these practices and replaces all occurrences of the
class usage with the updated syntax.

Disclaimer: The service locator pattern is present in several places
throughout ContentTranslation extension. The target of this patch is
not to eliminate all these cases, but only refactors this specific
class, so that we can start removing such bad practices step-by-step.

Change-Id: Icdd0f8d1c081101142262a68cf11dccd1941f411