SX recent translation entrypoint: Fix redirection to SX
Currently, when the user clicks the "Translate" button, they are
redirected to Section Translation, with the source title and the
language pair prefilled. However, instead of the title of the source
page, it is the title of the target page that is being used for this
redirection.

Fix this by using the source page title when redirecting to SX.

Bug: T287236
Change-Id: Icc3a9f0e9b441726ee35d37c03597f3a2f4b75dd