SX: Instrumentation support for all entrypoints
SX application properly supports the logging for the "dashboard_open"
event, with the "event_source" parameter indicating whether the user
is coming to SX from an entrypoint to the application. However, until
now only events for "mflanguagesearcher" were logged.

Fix this by adding a switch statement to also support the rest of the
entrypoints.

Bug: T301222
Change-Id: I6698b72aa1a444caaaeac599f36eb593c90cbb14