Fix: Template edits are not restored when translation resumed
Template parameter edits were not triggering a section change.
This was causing the edits lost when translation is saved and
then restored.

Process the transaction changes and find the corresponding sections
to register change.

Still, the restored templates are not rendered as per the target wiki
on translation restore. That is fixed in a follow up commit

Bug: T278323
Change-Id: I21a801a7128ce7303f474bab3ff75bb45ff804f2