CX Dependency Injection: add Wikibase SiteLinkLookup service
This patch registers Wikibase SiteLinkLookup as a service in
ContentTranslation extension, to be used for the recent edit
entrypoint.

This service already exists in service container when Wikibase
extension is installed, however, since this is an optional service
for the CX extension, this patch adds it as a nullable service
when Wikibase extension is not installed.

Change-Id: I9a028a257b0d6f5dbf4286162b0a01b01469874e
Bug: T305691