CX Dependency Injection: add Wikibase SettingsArray service
This patch registers Wikibase SettingsArray as a service in
ContentTranslation extension, to be used for the recent edit
entrypoint.

This service already exists in service container when Wikibase
extension is installed, however, since this is an optional service
for the CX extension, this patch adds it as a nullable service
when Wikibase extension is not installed.

Change-Id: I5f15b791c5366813da9bcd996e9a7aabe1c83aec
Bug: T305691