Allow translation by another user for existing translations after 24h
If there is an existing translation for selected language pair and title,
and if it is 24 hour old, allow it. Currently we don't allow translation
for such cases indefinitely.

This also fixes broken layout for translation conflict warning
That must be broken for long time. It uses the CX1 based layout.

Bug: T298244
Bug: T301218
Change-Id: Iae6d66b6687419b95f60bb13f51a01f1f5ad1b1e