tests: mw.cx.TranslationTracker: Fix for changes in VE
This test relied on HTML like `<section rel="cx:Section" …>` matching
ve.dm.SectionNode when ve.dm.CXSectionNode is not loaded. It will no
longer match after Id42967368224, where we've made it so that nodes
with unknown 'rel' attributes are consistently alienated.

Change it to just `<section …>` to preserve that behavior.

I also tried to fix it by loading ve.dm.CXSectionNode (by adding a
dependency on 'mw.cx.visualEditor'), but that causes the code to crash
because ve.init.target is not defined.

Bug: T303298
Change-Id: I785be75f389136769084a8e14832a09e3a519fc1