SX: Fix "learn more" link inside MT warning/error panel
The "Learn more" link inside the panel that displays the MT warnings
and errors inside the publishing step is currently not working and it
wrongfully refreshes the page when clicked.

This patch fixes this issue by adding the CX documentation page about
the translation limits as href and "_blank" as target attribute, so
that the link always opens in a new tab and current work is not being
lost. The redirection happens to a specific language version if it is
available.

Bug: T305131
Change-Id: If7a3314b444f8719dd43fed49f6a14c8f5dedb3b