SX: Enable CX beta feature when landed on Special:CX with valid campaigns
Similar to CX, enable CX beta feature if SX is accessed with a valid
campaign parameter. This is almost the case now, but there was a small
logic error while checking anon users. If SX is accessed by a logged in
user and anon access is enabled in that wiki, the above mentioned CX
beta feature enablement was not happening. Fix this by adding extra check
if user is anon there.

Bug: T298029
Change-Id: I0a836f2861378f842096c1b8cafa68909786313f