Avoid duplicate Category prefixes in published Translation
Due to a fix for older bug(T248302) there was a code to prefix
'Category:' if the restored category does not have a namespace prefix

User added categories will not have namesace prefixes and will get
always get this hardcoded 'Category:' prefix. When this is published to
a target language other than English, there will be an attempt to remove
the namespace prefix if any for the target language. Then a proper
category title will be constructed for publishing.

The hardcoded 'Category:' will never get removed and will get published
like "വർഗ്ഗം:Category:MycategoryName"

For fixing this issue:

1. Remove the 'Category:' prefixing code so that user added categories
   always saved and restored without any namesace prefixes.
2. As a preventive measure, in ApiContentTranslationPublish, along with
   removing category namespace prefix for target language, remove it for
   English as well to remove 'Category:' prefixing

Bug: T264490
Change-Id: Ife7202eea6292d30ccf2d0b9c814348161f256c5