Add invitation for translation when creating article in mobile
This patch adds a BeforePageDisplay hook handler that registers a JS
module, if the visited article doesn't exist, that adds an invitation
inside editor (both visual and wikitext mode) to start with a
translation instead of creating an article from scratch.

This JS module listens to the "mobileFrontend.editorOpened" event,
and it fetches suggestions for translation from cxserver, based on
the current language and page title. If at least one such suggestion
exists, the module creates the HTML for the invitation, before adding
it to the DOM.

Bug: T298084
Change-Id: I1debca9f17109e2f7a89d2d91deb0d1b64bc333d