SX Dashboard: Fix refresh button to work after multiple clicks
The refresh button for the suggestion lists inside SX Dashboard, only
works properly during the first 4 clicks. After that, the button
seems disabled, while it should display again the first suggestions
slice. This happens because according to the current implementation
the slice index is being updated, only when the next suggestions
slice has not been fetched (which is not true when the slice index
reaches the maximum number of slices that is hardcoded to 4).

Additionally, when a page suggestion already exists, before the
first page suggestions slice has been fetched (i.e. in case the
page is prefilled in the URL and the page is missing in target
language), the page suggestion refreshing works only once.

This patch fixes these issues by modifying the logic behind
refreshing, and it also applies some other fixes related to this
functionality:
* it removes the "fetchSectionSuggestionsBySeeds" and moves its
  functionality inside "fetchNextSectionSuggestionsSlice" action,
  as it was just adding a confusing redirection.
* adds the "getNumberOfPageSuggestionsToFetch" getter, similarly
  to the existing "getNumberOfSectionSuggestionsToFetch", so that
  already existing page suggestions are taken into account, when
  fetching new suggestions.
* adds a try/catch block around "fetchPageSuggestions" API method,
  so that the loading indicator is not eternally displayed when
  a request fails with not "ok" status.
* replaces some not needed, computed variables with regular ones,
  inside useSuggestions composable

Bug: T299891
Change-Id: I2e6739ecf70198dfcc89171edb2cbfd4a3052795