SX edit before publishing: Nest subsections into <section> elements
Currently, the content of the section to be translated is split
into subsections just before the "Pick a sentence" step. These
subsections are essentially the translation units that should be
stored inside parallel corpora table. However, these subsections
are merged again during the final edit inside "Preview and publish"
step, and thus the translation units cannot be properly published.

To maintain the structure of the translation units, this patch
nests the contents of each subsection into a "<section>" html
element. This way, the VE forces editing to happen for each
subsection separately, and the translation units are maintained
even after the final edit of the translation.

Additionally, this patch removes the "editedTranslation" property
inside the PageSection model and creates a new "editedTranslation"
property inside SubSection model. That means that the result of
the final edit of the section translation is stored piece-by-piece
inside each translated SubSection and the translation units can be
published properly.

Tests are fixed too.

Bug: T270499
Change-Id: Ia03e39d816c2174c4c306949c4dbb8d733508a4f