Sitemapper: Fix the configuration override
The $.extend function in jQuery modifies the first object passed.
Here it is a the config object from require( '../config.json' ).
This is a not a real file import, but a ResourceLoader package
file approach for importing. The config.json is mutable there.

Once config.json is modified, it applies to all later instantiation
of mw.cx.SiteMapper and hence we were getting TranslateInTarget
as false. This was leading to buggy redirects to same wiki for CX
when a new translation is started.

Fix the configuration override by passing the first object as {}.

Bug: T308802
Change-Id: I4d0c23ec944a102c66cd881cea9761ba256ce591