CX ULS entrypoint: Load only for content pages and Vector 2022 skin
Currently, the "ext.cx.uls.quick.actions" is basically loaded on all
pages when ContentTranslation extension is loaded. However, this
module is only needed when "Vector 2022" skin is used and the ULS is
available (that is only for content pages). Additionally, the module
should not be loaded for main pages or when the article action is not
"view" (e.g. "edit" or "history" mode).

This patch adds the appropriate conditions to only load the module
when actually needed.

Change-Id: I35581c30d1d8a35a40b95394603fc3c1c976123c