SX: Store the MT provider used for the applied translation
Until now, the MT provider that is used for each translated sentence
or block template is not stored. However, this information is
essential for the proper publishing of the translation units for
parallel corpora, so that the "origin" field is properly populated.

To properly support the "origin" property of the parallel corpora
translation units, this patch stores the MT provider used, each time
a proposed or edited translation is applied.

Bug: T270499
Change-Id: I38b30563e016999d5975ff999c1fa07aaeb91195