SX publishing: Always set applied section title as targetsectiontitle
Currently, the value of the "targetsectiontitle" property of the
publishing payload is calculated inside a getter that returns
different values depending on whether the translated is a lead
section, a present section and whether appendix sections exist inside
the target article. This approach worked well until now as the
"targetsectiontitle" parameter was only used as a "sectiontitle" for
the Edit API Action. However, in order to support the storage of
section translations the proper target section title should be
provided. This means that for all sections except for lead sections
the applied section title should be used as target section title, so
that the "cxsx_target_section_title" field inside
"cx_section_translations" table is populated correctly.

To support this need, this patch removes the getter that calculates
the target title for publishing and always use the applied section
title instead (or empty string for lead sections).

Change-Id: I5047c392b509a062c2c4ddd598fb87293f48aefd