SX subSection model: Add translationOrigin getter
Parallel corpora table contains a non-nullable "cxc_origin" field,
that contains the origin of the content, that is the MT provider
used for the corresponding translation unit.

In order to support that need, once the parallel corpora integration
is completed, this patch adds a "translationOrigin" getter inside
subSection model, that returns the origin for the translation unit
corresponding to the subSection model.

Since subSections inside SX are consisted by sectionSentences (in
case they are not block templates), multiple MT providers may be
used within a single subSection. For this reason, the getter returns
the origin that corresponds to the most used MT provider within the
subSection.

Bug: T270499
Change-Id: I678f084abb8ce5aba00bc2c9d3f57719d2f48f85