SX: Add request to "sxsave" api inside publishTranslation action
In order to support the parallel corpora integration for section
translations, the array of translation units should also be
published, along with the rest translation data, to the "sxsave"
API action. For this reason, this patch adds a call to the
"saveTranslation" translator API method, with the stringified
translation units array as "content" property inside the payload.

Additionally, this patch also adds the "sectionid" property inside
the payload. This property is basically an identifier for the
translated section in the folowing form:
"`source_revision_id`_`source_section_id`". Such an identifier is
essential for storing the section translation inside
"cx_section_translations" table.

Test added too.

Bug: T270499
Change-Id: I5cc88f3b167188e226430403f4d4e6e5a1b9fb3c