SX: Fix cleanup of unadapted cx-links to include all child nodes
Unadapted cx links that are not explicitly marked as missing should
be removed from the to-be-published HTML string. Until now these cx
links nodes were replaced from their innerHTML string. That works
well when the content of a cx-link node is a plain string, however
when these contents contain nested child nodes, the HTML string
for these nodes is being escaped, leading to weird text nodes,
instead of the original child nodes.

This patch fixes this issue by replacing the link with all the nested
child nodes, instead of its innerHTML. Tests are also added to verify
this behaviour.

Bug: T294007
Change-Id: I01ebbada058f09dfbdfc80fc3153a00e6795df4d