SX: Fix MT modification rate calculation
Until now, MT modification percentages are calculated wrongfully in
some cases. Calculating these modification percentages means
comparing the proposed MT translation with the applied translation
of a page section (or subsection). However, when the section contains
a translated block template, the proposed translation for this
template is not taken into account, leading to high MT modification
rates even if the proposed block template translation has not been
modified at all.

Additionally, another issue related to MT modification rates for
subsections (i.e. paragraphs) exist, when different MT providers have
been used for this paragraph.

This patch fixes those issues by removing the problematic logic that
was mentioned above. That is the MT for block templates is now added
properly inside the proposed translation used for the comparison, and
the proposed MT translation of a subsection(paragraph) is now
calculated based on the MT provider used, not the currently selected
one.

Bug: T309800
Change-Id: Iec7a43d08ba4ea4b6a9f1312707ebfd1aa738200