SX: Fix JS error for clicking on nested element inside sentence
Inside "Pick a sentence" step, when user tries to select a sentence,
and clicks on it, there is a chance that the target of the click
event is an element nested inside this sentence, like the <b>
element in the example below.

e.g. <span class="cx-segment"><b>Bold text</b>Other text</span>

Until now, in such cases, the translation unit to be selected, could
not be found, as the code only checked the event target itself, but
not its ancestors. That led to a JS error for the undefined
translation unit.

This patch fixes this issue by also checking the ancestors of the
clicked element, to find the sentence element and select it as
current translation unit.

Test added too.

Bug: T309331
Change-Id: Iae98ee416e66935321eaba0e36fd70a2e05da763