SX: Fix local resolution of reference definition
Articles usually have reference definition in reference list which
appears at end of article. The references within the paragraphs will
just have a pointer to those definitions. Since SX works with single
sections, the reference definition(or body) need to be resolved and
should be present in the section itself. We were using VE clipboard
mode for this resolution.

A recent change to the way nodes are parsed by
VE (Id4296736822439759c138223263ed44097d23cc6),
causes the CX Sections to not being identified as known node types.
The presence of rel attribute of <section> tags in CX caused VE to
treat it as unknown node. In SX, we had not registered this rel
attribute, but VE used to treat them as section nodes. The above
mentioned fix changed this behavior, and now it requires to register
the rel attribute with VE. Not registering them will cause VE to skip
all the nodes and the reference resolution will not work.

Register rel="cx:Section" with VE.

Bug: T309106
Change-Id: I66a41c2c0e8bfb76e9faa38df5f9080ecf9d18f0